backend/requirements.txt
fastapi
uvicorn
motor
openai
python-dotenv
pydantic
